0908lsj

+++ app/src/main/java/com/example/wantchu/ListStoreFavoritePage.java
@@ -109,11 +109,11 @@ public class ListStoreFavoritePage extends AppCompatActivity implements ListStor
             storeLocation.setLatitude(Double.parseDouble(favoriteListParsings.get(i).getStore_latitude()));
             storeLocation.setLongitude(Double.parseDouble(favoriteListParsings.get(i).getStore_longitude()));
 
-            double distance = getDistance(myLocation, storeLocation);
+            float distance = (float)getDistance(myLocation, storeLocation);
 
             Log.e("distance", String.valueOf(distance));
 
-            ListStoreHelperClass listStoreHelperClass = new ListStoreHelperClass(favoriteListParsings.get(i).getStore_name(),favoriteListParsings.get(i).getStore_location(),favoriteListParsings.get(i).getStore_image(),distance,favoriteListParsings.get(i).getStore_id(), favoriteListParsings.get(i).getStore_is_open());
+            ListStoreHelperClass listStoreHelperClass = new ListStoreHelperClass(favoriteListParsings.get(i).getStore_name(),favoriteListParsings.get(i).getStore_location(),favoriteListParsings.get(i).getStore_image(), distance,favoriteListParsings.get(i).getStore_id(), favoriteListParsings.get(i).getStore_is_open());
 
             listStoreHelperClass.storeNames.add(favoriteListParsings.get(i).getStore_name());
             listStoreHelperClass.storeLocations.add(favoriteListParsings.get(i).getStore_location());
@@ -152,7 +152,7 @@ public class ListStoreFavoritePage extends AppCompatActivity implements ListStor
             storeLocation.setLatitude(Double.parseDouble(favoriteListParsings.get(i).getStore_latitude()));
             storeLocation.setLongitude(Double.parseDouble(favoriteListParsings.get(i).getStore_longitude()));
 
-            double distance = getDistance(myLocation, storeLocation);
+            float distance = (float)getDistance(myLocation, storeLocation);
 
             Log.e("distance", String.valueOf(distance));
